chore: Ignore Git's FETCH_HEAD file
The FETCH_HEAD file is a temporary reference generated by `git fetch` to keep track of the branches just fetched from a remote repository.

This file is not part of the project's source code and should not be version-controlled. Add it to the .gitignore file to prevent it from being accidentally staged or committed in the future.